Create README.txt
Here is how to set up the project. 
First, download and install R studio. 
After downloading the files in this project, in R Studio in the menu go to Session > Set Working Directory > Choose Directory. Now choose the folder where
the downloaded files from this project are located. 
To run this project you will need to download the following packages to your R studio:
tidyverse, kableExtra, pairwiseCI, ggplot2.
To download a package:
In R Studio go to Tools > Install Packages. 
Once all the packages are downloaded and installed, to run the project in this repository, open HW2.Rmd in R Studio then use a keyboard shortcut Ctrl + Shift + K to knit the project.